third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbobjectstore_add5-expected.txt

This is a testharness.js-based test.
PASS IDBObjectStore.add() - object store's key path is an object attribute  
Harness: the test ran to completion.
